Make separate instance file for TxService
