+++ src/app/txServiceInstance.ts
@@ -1,0 +1,3 @@
+import TxService from "./TxServiceClass.ts";
+
+export default new TxService();
